Isolating the renderization process
