+++ src/routes/graph/RenderGraph.svelte
@@ -1,0 +1,232 @@
+<script lang="ts">
+	import { onMount } from 'svelte';
+	import { DataSet } from 'vis-data';
+	import { Network, type Node, type Edge } from 'vis-network';
+	import type Graph from 'graphology';
+	import { nip19 } from 'nostr-tools';
+	import { toasts } from 'svelte-toasts';
+
+	const renderData = {
+		nodes: new DataSet<Node>(),
+		edges: new DataSet<Edge>()
+	};
+
+	const physics = {
+		enabled: true,
+		barnesHut: {
+			theta: 0,
+			gravitationalConstant: -1000,
+			centralGravity: 0.05,
+			springLength: 80,
+			springConstant: 3 * 10 ** -5,
+			damping: 0.1,
+			avoidOverlap: 1
+		},
+		solver: 'barnesHut'
+	};
+
+	export let graph: Graph<Node, Edge>;
+
+	interface RenderChangedData {
+		nodes?: string[];
+		edges?: string[];
+	}
+
+	interface RenderOptions {}
+
+	const defaultRenderOptions = {};
+
+	export function render(data?: RenderChangedData, options: RenderOptions = defaultRenderOptions) {
+		const nodes = data?.nodes || graph.nodes();
+		const edges = data?.edges || graph.edges();
+
+		if (nodes.length === 0 && edges.length === 0) {
+			renderData.nodes.clear();
+			renderData.edges.clear();
+			return;
+		}
+
+		nodes.forEach((node) => {
+			if (!graph.hasNode(node)) {
+				renderData.nodes.remove(node);
+				return;
+			}
+
+			const data = graph.getNodeAttributes(node);
+
+			renderData.nodes.update({
+				id: node,
+				...data
+			});
+		});
+
+		edges.forEach((edge) => {
+			if (!graph.hasEdge(edge)) {
+				renderData.edges.remove(edge);
+				return;
+			}
+
+			const data = graph.getEdgeAttributes(edge);
+
+			renderData.edges.update({
+				id: edge,
+				...data
+			});
+		});
+	}
+
+	interface HoverWidths {
+		width: number;
+		hoverWidth: number;
+	}
+
+	export let nodeWidths: HoverWidths;
+
+	export let edgeWidths: HoverWidths;
+
+	export let physicsEnabled: boolean;
+
+	let graphContainer: HTMLDivElement;
+
+	let network: Network | undefined;
+
+	interface TogglePhysicsParams {
+		physicsEnabled: boolean;
+	}
+
+	function togglePhysics({ physicsEnabled }: TogglePhysicsParams) {
+		network?.setOptions({
+			physics: physicsEnabled ? physics : false
+		});
+	}
+
+	$: togglePhysics({ physicsEnabled });
+
+	onMount(() => {
+		network = new Network(graphContainer, renderData, {
+			physics: physicsEnabled ? physics : false,
+			nodes: {
+				shape: 'circularImage',
+				font: {
+					size: 40
+				},
+				brokenImage: '/avatar.svg',
+				chosen: false,
+				color: '#6b7891',
+				borderWidth: nodeWidths.width
+			},
+			groups: {
+				principal: {
+					size: 128
+				},
+				common: {
+					size: 64
+				}
+			},
+			edges: {
+				width: edgeWidths.width,
+				arrows: {
+					to: true
+				},
+				chosen: false,
+				smooth: {
+					type: 'continuous',
+					enabled: true,
+					forceDirection: 'none',
+					roundness: 0.1
+				}
+			},
+			interaction: { hover: true }
+		});
+
+		network.on('hoverNode', (event) => {
+			const nodeId = event.node as string;
+
+			renderData.nodes.update({
+				id: nodeId,
+				borderWidth: nodeWidths.hoverWidth
+			});
+		});
+
+		network.on('blurNode', (event) => {
+			const nodeId = event.node as string;
+
+			renderData.nodes.update({
+				id: nodeId,
+				borderWidth: nodeWidths.width
+			});
+		});
+
+		network.on('hoverEdge', (event) => {
+			const edgeId = event.edge as string;
+
+			renderData.edges.update({
+				id: edgeId,
+				width: edgeWidths.hoverWidth
+			});
+		});
+
+		network.on('blurEdge', (event) => {
+			const edgeId = event.edge as string;
+
+			renderData.edges.update({
+				id: edgeId,
+				width: edgeWidths.width
+			});
+		});
+
+		network.on('click', async (event) => {
+			const nodes = event.nodes as string[];
+
+			const edges = event.edges as string[];
+
+			function getActionType() {
+				if (nodes.length === 1) return 'pubkey';
+
+				if (edges.length === 1) return 'rating';
+			}
+
+			const actionType = getActionType();
+
+			if (!actionType) return;
+
+			const actions = {
+				pubkey: async () => {
+					const [node] = nodes;
+
+					const npub = nip19.npubEncode(node);
+
+					await navigator.clipboard.writeText(npub);
+
+					toasts.success({
+						title: 'Copied!',
+						description: 'NPUB copied to clipboard!'
+					});
+				},
+				rating: async () => {
+					const [edge] = edges;
+
+					const edgeData = graph.getEdgeAttributes(edge);
+
+					const fromNpub = nip19.npubEncode(edgeData.from as string);
+
+					const toNpub = nip19.npubEncode(edgeData.to as string);
+
+					const url = new URL('/table', window.location.origin);
+
+					url.searchParams.set('rater', fromNpub);
+
+					url.searchParams.set('rated', toNpub);
+
+					window.open(url.toString(), '_blank');
+				}
+			} as Record<typeof actionType, () => Promise<void>>;
+
+			await actions[actionType]?.();
+		});
+
+		network.moveTo({ position: { x: 0, y: 0 }, scale: 0.5 });
+	});
+</script>
+
+<div bind:this={graphContainer} class="w-full flex-1 bg-slate-400" />
